ai_ceo_platform/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
groq==0.4.2
requests==2.31.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
httpx==0.25.2
